ref(commands): refactored comamands
