+++ src/config/lib/helpers/сommandsWrapper.ts
@@ -1,14 +1,17 @@
 import { bot } from 'bot';
 import i18next, { t } from 'i18next';
 import db from 'config/db/databaseServise';
-import keyboard from 'bot/keyboard';
 import { getUserLanguage } from 'config/lib/helpers/cacheLaguage';
-import { notRegistrationMessage } from 'config/lib/helpers/notRegistrationMessage';
 import { sendMessage } from 'config/lib/helpers/sendMessage';
+import { notRegistrationMessage } from './notRegistrationMessage';
+import { type ICommandHandler } from 'config/types';
 
-export default (): void => {
-  const regex = /Отслеживать|Адсочваць/;
-  bot.onText(regex, (ctx) => {
+export async function сommandsWrapper({
+  regex,
+  handler,
+  options = { public: false },
+}: ICommandHandler): Promise<void> {
+  bot.onText(regex, (ctx, match) => {
     void (async () => {
       const userId = ctx.chat.id;
       await i18next.changeLanguage(await getUserLanguage(userId));
@@ -20,14 +23,14 @@ export default (): void => {
         );
         return;
       }
-      const isRegistered = await db.getUser(userId);
-      if (isRegistered) {
-        await sendMessage(
-          userId,
-          t('Сообщение об отслеживании'),
-          keyboard.Observe(),
-        );
-      } else await notRegistrationMessage(userId);
+      if (!options.public) {
+        const isRegistered = await db.getUser(userId);
+        if (!isRegistered) {
+          await notRegistrationMessage(userId);
+          return;
+        }
+      }
+      await handler(userId, match);
     })();
   });
-};
+}
